refactor: 同時に使う実装を同じファイルにまとめる
