--- apps/app/lib/router/navigator/setting_navigator.dart
@@ -1,12 +1,0 @@
-import 'package:features_setting/ui.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_app/router/provider/router.dart';
-
-final class SettingNavigatorImpl implements SettingNavigator {
-  const SettingNavigatorImpl();
-
-  @override
-  void goLicensePage(BuildContext context) {
-    const LicensePageRoute().go(context);
-  }
-}
