+++ apps/app/lib/router/provider/router.dart
@@ -8,9 +8,6 @@ import 'package:features_webview/webview.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/main_page.dart';
-import 'package:flutter_app/router/navigator/github_repository_list_navigator.dart';
-import 'package:flutter_app/router/navigator/home_navigator.dart';
-import 'package:flutter_app/router/navigator/setting_navigator.dart';
 import 'package:flutter_app/ui/home_page.dart';
 import 'package:flutter_app/ui/provider/navigator_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
